requirements.txt
matplotlib==3.10.1
numpy==2.2.3
Pillow==11.1.0
pygame==2.6.1
